Inverted the previous and next buttons
It looked to me that the buttons were inverted. It might be the case that it's a ltr-rtl issue?
I'm not sure.

However, I put the *previous post* button to the left and the *next post* button to the right.

Take a look and let me know if this fixes the issue or if it's just a misconfiguration on my end.


Former-commit-id: ab191d3185356fc3c091151cf014a130c901416c [formerly 79c4c59df0e61f778050e75207b660cbb364cde0]
Former-commit-id: 34c5a506b2b6abe371af09a73155ffc6514f10b0