remove quote from start of post
Since the post was published, I've been contacted by someone who made me aware that the wording "found my tribe" can be offensive to some oppressed groups e.g. indigenous peoples. I'd like to remove the quote from the top of the blogpost and add a note to the bottom of the post, to explain the edit.