Remove broken link and suggestions for replacement
The Medium essay shared under the section on demographics has been deleted by the author, so the link should be removed.

It might be nice to have a replacement link. In lieu of a specific essay, what about linking to a library blog or journal? For example, Library Journal's "People" section (https://www.libraryjournal.com/section/people) shares Q&As with librarians in all sorts of jobs, in addition to sharing editorials on topics such racial equity, book challenges, representation, and the digital divide. In the Library with the Lead Pipe (https://www.inthelibrarywiththeleadpipe.org/) is a journal that often publishes articles relating to diversity, equity, and inclusion. While neither resource is fully centered on demographic information, they could give insight into the library field and what librarians are seeing in their libraries and workplaces.